refactor: extract checkWhitespaceIssues function and TextAreaModal
- Extract checkWhitespaceIssues to a separated file
- Extract TextAreaModal to a separated file
- Make Warning Icon clickable and allow it open TextAreaModal
- Add little scale on icon button on hover/click in order to provide a clue to the user that it is clickable
