+++ frontend/web/styles/project/_forms.scss
@@ -123,6 +123,19 @@
     font-weight: 500;
 }
 
+.rule-value-icon {
+    transition: all 0.2s ease-in-out;
+
+    &:hover {
+        transform: scale(1.05);
+    }
+
+    &:active {
+        transform: scale(0.95);
+        transition: all 0.1s ease-in-out;
+    }
+}
+
 .setting {
     label {
         font-size: $font-size-base;
